+++ docs/.vitepress/config.ts
@@ -62,7 +62,9 @@ export default defineConfig({
       {
         text: 'Components',
         items: [
+          { text: 'Alert', link: '/components/alert' },
           { text: 'Button', link: '/components/button' },
+          { text: 'Card', link: '/components/card' },
           { text: 'Checkbox', link: '/components/checkbox' },
           { text: 'Form Item', link: '/components/form-item' },
           { text: 'Input', link: '/components/input' },